fix migrations

+++ database/migrations/2022_07_27_101721_create_deliveries_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('deliveries', function (Blueprint $table) {
             $table->id();
             $table->string("name");
-            $table->unsignedInteger('id_user');
+            $table->unsignedBigInteger('id_user');
             $table->foreign('id_user')->references('id')->on('users')->onDelete('cascade');
             $table->string('address');
             $table->string('zip');
